requirements.txt
streamlit
sqlalchemy
pandas
plotly
python-docx